+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) from version 0.3.0 forward.
 
+## Unreleased
+
+### Added
+
+- stats.StochasticDataComparison(): added daily correlation outputs
+- stats.StochasticDataComparison(): added monthly statistic charts grouped by timeseries
+- stats.StochasticDataComparison(): added optional 'show_bands' argument to show range of outcomes as grey band on statistic charts
+
+### Fixes
+
+- stats.StochasticDataComparison(): no longer ignores wy_month argument in cropping input datasets
+
+### Changed
+
+- stats.StochasticDataComparison(): legend in charts is now interactive, subset of input datasets can now be shown
+- stats.StochasticDataComparison(): adjusted correlation charts so that input dataset colour is consistent with other charts
+- stats.StochasticDataComparison(): sort order of all chart colours now matches input dataset order
+
 ## 0.3.0
 
 ### Added